Allow managing web subscriptions from native
