+++ src/state/purchases/hooks/useManageSubscription/index.ts
@@ -2,15 +2,23 @@ import {Linking} from 'react-native'
 import Purchases from 'react-native-purchases'
 import {useMutation} from '@tanstack/react-query'
 
-export function useManageSubscription() {
+import {useManageSubscription as useManageWebSubscription} from '#/state/purchases/hooks/useManageSubscription/web'
+import {PlatformId} from '#/state/purchases/types'
+
+export function useManageSubscription({platform}: {platform: PlatformId}) {
+  const web = useManageWebSubscription({platform})
   return useMutation({
     async mutationFn() {
-      const customer = await Purchases.getCustomerInfo()
-
-      if (customer.managementURL) {
-        Linking.openURL(customer.managementURL)
+      if (platform === PlatformId.Web) {
+        await web.mutateAsync()
       } else {
-        // TODO
+        const customer = await Purchases.getCustomerInfo()
+
+        if (customer.managementURL) {
+          Linking.openURL(customer.managementURL)
+        } else {
+          // TODO
+        }
       }
     },
   })
